# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API de Provincias - TP08 DAI",
    "description": "API propia (Node.js + Express + PostgreSQL) para la gestión de un catálogo de provincias: alta, baja, modificación total/parcial y consulta.",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/provinces/": {
      "get": {
        "tags": [
          "Provinces"
        ],
        "description": "Obtiene todas las provincias registradas.",
        "responses": {
          "200": {
            "description": "Listado de provincias",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Province"
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "post": {
        "tags": [
          "Provinces"
        ],
        "description": "Crea una nueva provincia. Requiere al menos \"name\" (mínimo 3 caracteres).",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Datos de la provincia a crear",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProvinceInput"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Provincia creada",
            "schema": {
              "$ref": "#/definitions/Province"
            }
          },
          "400": {
            "description": "Nombre faltante o con menos de 3 caracteres"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "put": {
        "tags": [
          "Provinces"
        ],
        "description": "Reemplaza una provincia existente por completo. El id se envía dentro del body (no en la ruta).",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Provincia completa, incluyendo el id",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Province"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provincia actualizada",
            "schema": {
              "$ref": "#/definitions/Province"
            }
          },
          "400": {
            "description": "Nombre faltante o con menos de 3 caracteres"
          },
          "404": {
            "description": "Provincia no encontrada"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/api/provinces/{id}": {
      "get": {
        "tags": [
          "Provinces"
        ],
        "description": "Obtiene una provincia específica según su id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID de la provincia"
          }
        ],
        "responses": {
          "200": {
            "description": "Provincia encontrada",
            "schema": {
              "$ref": "#/definitions/Province"
            }
          },
          "404": {
            "description": "No se encontró la provincia"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "patch": {
        "tags": [
          "Provinces"
        ],
        "description": "Actualiza parcialmente una provincia existente: solo modifica los campos enviados en el body.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID de la provincia"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Campos a modificar (no es necesario enviarlos todos)",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProvincePartial"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provincia actualizada",
            "schema": {
              "$ref": "#/definitions/Province"
            }
          },
          "400": {
            "description": "Nombre inválido (si se envía y no cumple la validación)"
          },
          "404": {
            "description": "Provincia no encontrada"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      },
      "delete": {
        "tags": [
          "Provinces"
        ],
        "description": "Elimina una provincia según su id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID de la provincia"
          }
        ],
        "responses": {
          "200": {
            "description": "Provincia eliminada"
          },
          "404": {
            "description": "Provincia no encontrada"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/api/provinces/bulk": {
      "post": {
        "tags": [
          "Provinces"
        ],
        "description": "Crea varias provincias en una sola operación, a partir de un array. Si alguna no pasa la validación, se corta y se devuelve ese error.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Array de provincias a crear",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProvinceInput"
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Provincias creadas",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Province"
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Se esperaba un array no vacío, o alguna provincia no pasó la validación"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/admin/reset/": {
      "post": {
        "tags": [
          "Admin"
        ],
        "description": "Resetea los datos de la tabla de provincias a su estado inicial (uso administrativo).",
        "responses": {
          "201": {
            "description": "Datos reseteados correctamente",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Province"
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Error en la request"
          }
        }
      }
    }
  },
  "definitions": {
    "Province": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "name": {
          "type": "string",
          "example": "Buenos Aires"
        },
        "full_name": {
          "type": "string",
          "example": "Provincia de Buenos Aires"
        },
        "latitude": {
          "type": "number",
          "example": -34.6
        },
        "longitude": {
          "type": "number",
          "example": -58.4
        },
        "display_order": {
          "type": "number",
          "example": 1
        }
      }
    },
    "ProvinceInput": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Buenos Aires"
        },
        "full_name": {
          "type": "string",
          "example": "Provincia de Buenos Aires"
        },
        "latitude": {
          "type": "number",
          "example": -34.6
        },
        "longitude": {
          "type": "number",
          "example": -58.4
        },
        "display_order": {
          "type": "number",
          "example": 1
        }
      }
    },
    "ProvincePartial": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Buenos Aires"
        }
      }
    }
  }
}